Changed S3 behaviour and verified Products, Assets and Tags linking

+++ backend/src/Services/productTags.service.ts
@@ -1,4 +1,5 @@
 import { prisma } from "../Configs/prisma";
+import { AppError } from "../Utilities";
 import { ProductService } from "./product.service";
 
 export class ProductTagsService {
@@ -19,6 +20,29 @@ export class ProductTagsService {
         return tags;
     }
     static async linkTags(productId: string, tagIds: string[]) {
+        const productExists = await prisma.products.findUnique({
+            where: { id: productId },
+            select: { id: true },
+        });
+
+        if (!productExists) {
+            throw new AppError("Invalid product ID — product not found.", 404);
+        }
+
+        const existingTags = await prisma.tags.findMany({
+            where: { id: { in: tagIds } },
+            select: { id: true },
+        });
+
+        if (existingTags.length !== tagIds.length) {
+            const missing = tagIds.filter(
+                (id) => !existingTags.find((tag) => tag.id === id)
+            );
+            throw new AppError(
+                `Some tags not found: ${missing.join(", ")}`,
+                400
+            );
+        }
         await prisma.product_tags.createMany({
             data: tagIds.map((tagId) => ({
                 product_id: productId,
@@ -31,6 +55,23 @@ export class ProductTagsService {
     }
 
     static async unlink(productId: string, tagId: string) {
+        const productExists = await prisma.products.findUnique({
+            where: { id: productId },
+            select: { id: true },
+        });
+
+        if (!productExists) {
+            throw new AppError("Invalid product ID — product not found.", 404);
+        }
+
+        const existingTag = await prisma.tags.findUnique({
+            where: { id: tagId },
+            select: { id: true },
+        });
+
+        if (!existingTag) {
+            throw new AppError(`Tag not found`, 400);
+        }
         await prisma.product_tags.delete({
             where: {
                 product_id_tag_id: {
